chore: use the new way of trpc
